+++ Rdmp.UI/CommandExecution/AtomicCommands/ExecuteCommandOpenExtractionDirectory.cs
@@ -7,6 +7,7 @@
 using System;
 using System.Drawing;
 using System.IO;
+using System.Linq;
 using Rdmp.Core.CommandExecution.AtomicCommands;
 using Rdmp.Core.DataExport.Data;
 using Rdmp.Core.Icons.IconProvision;
@@ -18,6 +19,7 @@ namespace Rdmp.UI.CommandExecution.AtomicCommands
     internal class ExecuteCommandOpenExtractionDirectory : BasicUICommandExecution,IAtomicCommand
     {
         private FileInfo _file;
+        private DirectoryInfo _dir;
 
         public ExecuteCommandOpenExtractionDirectory(IActivateItems activator, ISelectedDataSets sds):base(activator)
         {
@@ -26,15 +28,52 @@ namespace Rdmp.UI.CommandExecution.AtomicCommands
             {
                 if(result == null)
                     SetImpossible("Dataset has not been extracted");
-                else 
-                if (result.DestinationType == null || !result.DestinationType.EndsWith("FlatFileDestination"))
-                    SetImpossible("Extraction destination was not to disk");
+                else
+                if(result.DestinationType == null)
+                    SetImpossible("This extraction has not been run");
+                else if (!result.DestinationType.EndsWith("FlatFileDestination"))
+                    SetImpossible($"Extraction destination was '{result.DestinationType}' so cannot be opened");
                 else
                 {
                     _file = new FileInfo(result.DestinationDescription);
 
                     if(!_file.Exists)
                         SetImpossible($"File '{_file.FullName}' did not exist on disk");
+                }
+            }
+            catch (Exception)
+            {
+                SetImpossible("Could not determine file location");
+            }
+
+        }
+
+        public ExecuteCommandOpenExtractionDirectory(IActivateItems activator, IExtractionConfiguration configuration) : base(activator)
+        {
+            var cumulativeExtractionResults = configuration.SelectedDataSets.Select(s=>s.GetCumulativeExtractionResultsIfAny()).Where(c=>c!=null).ToArray();
+            try
+            {
+                if (cumulativeExtractionResults.Length == 0)
+                    SetImpossible("No datasets have ever been extracted");
+                else
+                if (!cumulativeExtractionResults.All(c=>c.DestinationType != null && c.DestinationType.EndsWith("FlatFileDestination")))
+                    SetImpossible("Extraction destinations were not to disk");
+                else
+                {
+                    // all datasets have been extracted to disk
+
+                    // but do they have a shared parent dir?
+                    var files = cumulativeExtractionResults.Select(c => new FileInfo(c.DestinationDescription)).ToArray();
+
+                    var parents = files.Select(f => f.Directory?.Parent?.FullName).Where(d=>d != null).Distinct().ToArray();
+
+                    if (parents.Length != 1)
+                        SetImpossible($"Extracted files do not share a common extraction directory");
+                    else
+                    {
+                        _dir = new DirectoryInfo(parents[0]);
+                    }
+                        
                 }
             }
             catch (Exception)
@@ -53,7 +92,9 @@ namespace Rdmp.UI.CommandExecution.AtomicCommands
         {
             base.Execute();
 
-            var cmd = new ExecuteCommandOpenInExplorer(Activator, _file);
+            var cmd = _file != null?
+                new ExecuteCommandOpenInExplorer(Activator, _file):
+                new ExecuteCommandOpenInExplorer(Activator, _dir);
             
             if(!cmd.IsImpossible)
                 cmd.Execute();